move binstream from source_io to module_out in source_base
