+++ source/source_base/CMakeLists.txt
@@ -59,6 +59,7 @@ add_library(
     cubic_spline.cpp
     parallel_2d.cpp
     projgen.cpp
+    module_out/binstream.cpp
     module_mixing/mixing_data.cpp
     module_mixing/mixing.cpp
     module_mixing/plain_mixing.cpp
