+++ source/source_io/CMakeLists.txt
@@ -26,7 +26,6 @@ list(APPEND objects
     module_wf/read_wfc_pw.cpp
     module_wf/read_wf2rho_pw.cpp
     module_restart/restart.cpp
-    module_output/binstream.cpp
     module_wf/write_wfc_pw.cpp
     module_output/write_pao.cpp
     module_output/write_cube.cpp
