+++ source/source_io/module_chgpot/rhog_io.cpp
@@ -1,4 +1,4 @@
-#include "source_io/module_output/binstream.h"
+#include "source_base/module_out/binstream.h"
 #include "source_base/global_function.h"
 #include "source_io/module_parameter/parameter.h"
 #include "source_base/global_variable.h"
