+++ source/source_io/module_wf/write_wfc_pw.cpp
@@ -4,7 +4,7 @@
 #include "mpi.h"
 #endif
 
-#include "source_io/module_output/binstream.h"
+#include "source_base/module_out/binstream.h"
 #include "source_base/global_variable.h"
 #include "source_base/parallel_global.h"
 #include "source_base/tool_title.h"
