+++ source/source_io/test/write_wfc_nao_test.cpp
@@ -4,7 +4,7 @@
 #define private public
 #include "source_io/module_parameter/parameter.h"
 #undef private
-#include "../module_output/binstream.h"
+#include "source_base/module_out/binstream.h"
 #include "source_base/global_variable.h"
 #include "source_base/module_external/scalapack_connector.h"
